Imported Dev branch and moved classes to DBManagement Package
