+++ SpartiosDatabaseManager/src/main/java/com/sparta/spartios/dbmanagement/DatabaseCreationStatements.java
@@ -1,4 +1,4 @@
-package com.sparta.spartios;
+package com.sparta.spartios.dbmanagement;
 
 public interface DatabaseCreationStatements {
